requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
celerybeat-redis==0.1.5
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.15
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-celerybeat-status==0.0.10
django-redis==4.12.1
django-rest-swagger==2.2.0
django-timezone-field==4.0
djangorestframework==3.11.0
idna==2.10
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.11
MarkupSafe==1.1.1
minimalmodbus==1.0.2
mysqlclient==2.0.1
openapi-codec==1.3.2
pyserial==3.4
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
zipp==1.2.0
